Fix incorrect casing when rendering fields amd enums.
This change fixed incorrect case rendering of fields inside a record
which has the same record type and uses a reference of the type to
determine the type of the field. Since the field name is converted to
Pascal case for Record types, the reference type name should also be
Pascal cased.
The Enum class rendering has a similar issue where the default value
of the enum field does not contain the Pascal case of the Enum class.