fix: incorrect casing when rendering fields and enums  (#694). Closes #693 and #692
* Fix incorrect casing when rendering fields amd enums.

This change fixed incorrect case rendering of fields inside a record
which has the same record type and uses a reference of the type to
determine the type of the field. Since the field name is converted to
Pascal case for Record types, the reference type name should also be
Pascal cased.
The Enum class rendering has a similar issue where the default value
of the enum field does not contain the Pascal case of the Enum class.

* Render meta class for record and enum types

Render a meta class with a schema_name attribute anytime when the type name
is changed to pascal case in python classes. This will enable to obtain the
original schema back from python classes.
Updated unit tests.

* Fix linting issues

Fix lines too long issue

* Add changes related to PR feedback

---------

Co-authored-by: Marcos Schroh <2828842+marcosschroh@users.noreply.github.com>
Co-authored-by: Sanjay Jayaramu <sanjayjayaramu@gmail.com>