feat(AvroBaseModel): non-deprecated pydantic v2 custom fields
Add a new way to generate Avro schema and serialize custom fields for Pydantic v2 `AvroBaseModel`s. This new strategy provides a `CustomAvroEncoder` class to use as an annotation on fields with types that don't natively support serialization to Avro. The existing Pydantic v2 custom serialization/schema-generation strategy depends on the `json_encoders` model config, which is deprecated.