fix(TimedeltaField): fix timedelta fastavro serialization so it doesn't break in union types (#797)
Fastavro [custom logical type encoders](https://fastavro.readthedocs.io/en/latest/logical_types.html#custom-logical-types) should explicitly check that they are operating on the expected python type before encoding it, and return the unchanged data otherwise. If a custom logical type is used as a member of a union type, then this encoder function will be called on any value that is provided for a field of that type, even if the value's type is a different member of the union type.