fix: support Python 3.13 in is_annotated
In Python 3.13, the `typing.Annotated` type was changed to no longer
support `isinstance` or `issubclass`. Replace the `is_annotated`
implementation with code from Pydantic that checks to see if the
origin of the type matches the `Annotated` type exported from either
`typing` or `typing_extensions`.

Add Python 3.13 to the CI test matrix and to the library classifiers.

Fixes #816