Fixes the Pydantic's custom field-level attributes
The custom field-level documentation when using Pydantic is referencing
the same instructions as if you were using data classes, but this
doesn't work, as Pydantic doesn't have a `metadata` key in their `Field`
function.

After some debugging, I found that you have to use the `json_schema_extra`
parameter in `Field`, passing a dictionary whose main key is `metadata`,
with the object containing the field-level metadata you want to add.

I tested this out with the latest Pydantic version and it works. I am
unsure about Pydantic's v1, so this PR is only addressing the latest
Pydantic docs.

I am happy to update the documentation for v1 if it works the same way.